move expandOpenApiProps to ts helper file
